Make more progress towards PMEM
Working out how to recompile a function with the memory state of inner
nodes toggled. It would be nice to be able to mutate an existing
function, as the node names would be consistent for easier profiling.

There does not seem to be native support for copying functions, but we
could always serialize a function and then deserialize it to get the
copying behavior.