A less messed up way of doing PMEM
Add a `pool_number` member to `descriptor::Tensor`. This feels natural
since the notion of `pool_offset` already exists in the class.

Cleaned up some other code as well to make the implementation a bit
better.

Removed the `PERSISTENT_INTERMEDIATE` role from CPUTensorRole because it
felt wrong and is also no longer needed.